Ré-organisation avec séparation de la logiqu métier
