+++ dags/sources/source_sinoe.py
@@ -2,8 +2,8 @@ import os
 
 import pandas as pd
 from airflow import DAG
+from sources.tasks.airflow_logic.operators import default_args, eo_task_chain
 from utils.base_utils import get_mapping_config, source_sinoe_dechet_mapping_get
-from utils.eo_operators import default_args, eo_task_chain
 
 # ------------------------------------------
 # PARAMÈTRES ET DOCUMENTATION DU DAG
