+++ dags/sources/source_soren.py
@@ -1,6 +1,6 @@
 from airflow import DAG
+from sources.tasks.airflow_logic.operators import default_args, eo_task_chain
 from utils.base_utils import get_mapping_config
-from utils.eo_operators import default_args, eo_task_chain
 
 with DAG(
     dag_id="eo-soren",
