+++ dags/sources/tasks/airflow_logic/operators.py
@@ -1,0 +1,116 @@
+from datetime import datetime
+
+from airflow import DAG
+from airflow.models.baseoperator import chain
+from airflow.operators.python import PythonOperator
+from sources.tasks.airflow_logic.db_read_acteur_task import db_read_acteur_task
+from sources.tasks.airflow_logic.db_read_propositions_max_id_task import (
+    db_read_propositions_max_id_task,
+)
+from sources.tasks.airflow_logic.get_acteur_to_delete_task import (
+    propose_acteur_to_delete_task,
+)
+from sources.tasks.airflow_logic.propose_acteur_changes_task import (
+    propose_acteur_changes_task,
+)
+from sources.tasks.airflow_logic.propose_acteur_services_task import (
+    propose_acteur_services_task,
+)
+from sources.tasks.airflow_logic.propose_labels_task import propose_labels_task
+from sources.tasks.airflow_logic.propose_services_sous_categories_task import (
+    propose_services_sous_categories_task,
+)
+from sources.tasks.airflow_logic.propose_services_task import propose_services_task
+from sources.tasks.airflow_logic.read_mapping_from_postgres_task import (
+    read_mapping_from_postgres_task,
+)
+from sources.tasks.airflow_logic.source_config_validate_task import (
+    source_config_validate_task,
+)
+from sources.tasks.airflow_logic.source_data_download_task import (
+    source_data_download_task,
+)
+from sources.tasks.airflow_logic.source_data_normalize_task import (
+    source_data_normalize_task,
+)
+from sources.tasks.airflow_logic.source_data_validate_task import (
+    source_data_validate_task,
+)
+from utils import dag_eo_utils
+
+default_args = {
+    "owner": "airflow",
+    "depends_on_past": False,
+    "start_date": datetime(2024, 3, 23),
+    "email_on_failure": False,
+    "email_on_retry": False,
+    "retries": 1,
+}
+
+
+# TODO : faire un wrapper & déplacer dans un fichier dédié
+def write_data_task(dag: DAG) -> PythonOperator:
+    return PythonOperator(
+        task_id="db_data_write",
+        python_callable=dag_eo_utils.write_to_dagruns,
+        dag=dag,
+    )
+
+
+# TODO : faire un wrapper & déplacer dans un fichier dédié
+def serialize_to_json_task(dag: DAG) -> PythonOperator:
+    return PythonOperator(
+        task_id="db_data_prepare",
+        python_callable=dag_eo_utils.db_data_prepare,
+        dag=dag,
+    )
+
+
+def eo_task_chain(dag: DAG) -> None:
+    read_tasks = [
+        read_mapping_from_postgres_task(
+            dag=dag, table_name="qfdmo_acteurtype", task_id="db_read_acteurtype"
+        ),
+        read_mapping_from_postgres_task(
+            dag=dag, table_name="qfdmo_source", task_id="db_read_source"
+        ),
+        read_mapping_from_postgres_task(
+            dag=dag, table_name="qfdmo_action", task_id="db_read_action"
+        ),
+        read_mapping_from_postgres_task(
+            dag=dag, table_name="qfdmo_acteurservice", task_id="db_read_acteurservice"
+        ),
+        read_mapping_from_postgres_task(
+            dag=dag, table_name="qfdmo_labelqualite", task_id="db_read_labelqualite"
+        ),
+        read_mapping_from_postgres_task(
+            dag=dag,
+            table_name="qfdmo_souscategorieobjet",
+            task_id="db_read_souscategorieobjet",
+        ),
+        db_read_propositions_max_id_task(dag),
+    ]
+
+    create_tasks = [
+        propose_services_task(dag),
+        propose_labels_task(dag),
+        propose_acteur_services_task(dag),
+    ]
+
+    chain(
+        # Spécifique à la source pour amener à
+        # un état normalisé et validé des données
+        source_config_validate_task(dag),
+        source_data_download_task(dag),
+        source_data_normalize_task(dag),
+        source_data_validate_task(dag),
+        # Logique commune à toutes les sources
+        read_tasks,
+        db_read_acteur_task(dag),
+        propose_acteur_changes_task(dag),
+        propose_acteur_to_delete_task(dag),
+        create_tasks,
+        propose_services_sous_categories_task(dag),
+        serialize_to_json_task(dag),
+        write_data_task(dag),
+    )
