+++ dags/sources/tasks/airflow_logic/propose_acteur_services_task.py
@@ -1,0 +1,30 @@
+import logging
+
+from airflow import DAG
+from airflow.operators.python import PythonOperator
+from sources.tasks.business_logic.propose_acteur_services import propose_acteur_services
+from utils import logging_utils as log
+
+logger = logging.getLogger(__name__)
+
+
+# TODO : faire un wrapper
+def propose_acteur_services_task(dag: DAG) -> PythonOperator:
+    return PythonOperator(
+        task_id="propose_acteur_services",
+        python_callable=propose_acteur_services_wrapper,
+        dag=dag,
+    )
+
+
+def propose_acteur_services_wrapper(**kwargs):
+    acteurservice_id_by_code = kwargs["ti"].xcom_pull(task_ids="db_read_acteurservice")
+    df_actors = kwargs["ti"].xcom_pull(task_ids="propose_acteur_changes")["df"]
+
+    log.preview(df_actors, "df_actors")
+    log.preview(acteurservice_id_by_code, "acteurservice_id_by_code")
+
+    return propose_acteur_services(
+        df_actors=df_actors,
+        acteurservice_id_by_code=acteurservice_id_by_code,
+    )
