+++ dags/sources/tasks/business_logic/db_read_acteur.py
@@ -7,16 +7,6 @@ from utils import logging_utils as log
 logger = logging.getLogger(__name__)
 
 
-def db_read_acteur_wrapper(**kwargs):
-    df_normalized = kwargs["ti"].xcom_pull(task_ids="source_data_normalize")
-
-    log.preview("df_normalized", df_normalized)
-
-    return db_read_acteur(
-        df_normalized=df_normalized,
-    )
-
-
 def db_read_acteur(
     df_normalized: pd.DataFrame,
 ):
