+++ dags/sources/tasks/business_logic/source_data_validate.py
@@ -8,19 +8,6 @@ from utils import logging_utils as log
 logger = logging.getLogger(__name__)
 
 
-def source_data_validate_wrapper(**kwargs) -> None:
-    df = kwargs["ti"].xcom_pull(task_ids="source_data_normalize")
-    params = kwargs["params"]
-
-    log.preview("df depuis source_data_normalize", df)
-    log.preview("paramètres du DAG", params)
-
-    return source_data_validate(
-        df=df,
-        params=params,
-    )
-
-
 def source_data_validate(
     df: pd.DataFrame,
     params: dict,
